oaiharvester: Holding Pen formatting
* Moves the Holding Pen rendering functions to the OAI
  harvest workflow 'oaiharvest_harvest_repositories' in order
  to have more OAI harvest workflows profit from them.
