+++ invenio/modules/workflows/templates/workflows/styles/harvesting_description.html
@@ -23,10 +23,12 @@
 
   {% if result_progress and result_progress.total > 0 %}
 
+    {% set total_succeeded = result_progress.total - result_progress.failed %}
     <h5>{{ _('Progress') }}:</h5>
 
     <div class="progress">
-      <div class="progress-bar progress-bar-success" style="width: {{(result_progress.total - result_progress.failed)* 100 / result_progress.total}}%">
+      <div class="progress-bar progress-bar-success"
+           style="width: {{(total_succeeded)* 100 / result_progress.total}}%">
         <span class="sr-only">{{ _('Succeeded') }}</span>
       </div>
       <div class="progress-bar progress-bar-danger" style="width: {{result_progress.failed * 100 / result_progress.total}}%">
@@ -34,7 +36,8 @@
       </div>
     </div>
 
-    <strong>{{result_progress.total - result_progress.failed}}</strong> {{_("out of")}} <strong>{{result_progress.total}}</strong> {{_("record(s) done")}}.
+    <strong>{{total_succeeded}}</strong> {{_("out of")}}
+    <strong>{{result_progress.total}}</strong> {{_("record(s) done")}}.
 
     {% if result_progress.failed %}
       <p class="text-warning">
@@ -53,9 +56,15 @@
     <div class="list-group">
       {% for id_object, title in related_objects %}
         {% if not title %}
-          <a href="{{ url_for('holdingpen.details', objectid=id_object) }}" class="list-group-item"><span class="badge">{{id_object}}</span> {{ _('Record deleted') }}</a>
+          <a href="{{ url_for('holdingpen.details', objectid=id_object) }}"
+             class="list-group-item">
+             <span class="badge">{{id_object}}</span> {{ _('Record deleted') }}
+          </a>
         {% else %}
-          <a href="{{ url_for('holdingpen.details', objectid=id_object) }}" class="list-group-item"><span class="badge">{{id_object}}</span>{{title}}</a>
+          <a href="{{ url_for('holdingpen.details', objectid=id_object) }}"
+             class="list-group-item">
+             <span class="badge">{{id_object}}</span>{{title}}
+          </a>
         {% endif %}
       {% endfor %}
     </div>
